+++ jobs/fedsm_prostate/app/custom/medclip/losses.py
@@ -43,10 +43,9 @@ class ImageTextContrastiveLoss(nn.Module):
 
             # get logits
             logits = outputs['logits']
-
             # compute soft-labels, -1: negative, 0: uncertain, 1: positive
             # in the original data: 1: positive, 0: negative, -1: uncertain, NA: not mentioned
-            label_sim = torch.matmul(img_labels, text_labels.T)
+            label_sim = torch.matmul(img_labels.to(torch.float32), text_labels.T.to(torch.float32)).to(torch.bfloat16)
             label_sim = label_sim.to(logits.device)
 
             if aug_input_ids is not None:
